sparseFusion compiling again after changes to DPvMFmeans simple

+++ include/tdp/preproc/plane.h
@@ -24,7 +24,7 @@ struct Plane {
 
   Brief feat_;
 
-  uint32_t z_;
+  uint16_t z_;
   uint32_t lastFrame_;
   uint32_t numObs_;
